fix: #104 separate handle file modelApi and modelLocal
